[Fix](BE) : 정산 시 오버플로우 발생해서 int -> long으로 변경

+++ BE/mvp/src/main/java/com/mvp/stats/dto/ParkingLotSpotStats.java
@@ -18,6 +18,6 @@ public class ParkingLotSpotStats {
     private Integer carState;
     private LocalDateTime entranceTime;
     private LocalDateTime exitTime;
-    private Integer fee;
+    private Long fee;
     private byte[] image;
 }
